moved view helper to root
