+++ themes/root/theme.config.php
@@ -11,6 +11,7 @@ return [
             'VuFind\View\Helper\Root\AlphaBrowse' => 'VuFind\View\Helper\Root\AlphaBrowseFactory',
             'VuFind\View\Helper\Root\Auth' => 'VuFind\View\Helper\Root\AuthFactory',
             'VuFind\View\Helper\Root\AuthorNotes' => 'VuFind\View\Helper\Root\ContentLoaderFactory',
+            'VuFind\View\Helper\Root\AvailabilityStatus' => 'Laminas\ServiceManager\Factory\InvokableFactory',
             'VuFind\View\Helper\Root\Browse' => 'Laminas\ServiceManager\Factory\InvokableFactory',
             'VuFind\View\Helper\Root\BulkAction' => 'VuFind\View\Helper\Root\BulkActionFactory',
             'VuFind\View\Helper\Root\Captcha' => 'VuFind\View\Helper\Root\CaptchaFactory',
@@ -103,6 +104,7 @@ return [
             'alphabrowse' => 'VuFind\View\Helper\Root\AlphaBrowse',
             'auth' => 'VuFind\View\Helper\Root\Auth',
             'authorNotes' => 'VuFind\View\Helper\Root\AuthorNotes',
+            'availabilityStatus' => 'VuFind\View\Helper\Root\AvailabilityStatus',
             'browse' => 'VuFind\View\Helper\Root\Browse',
             'bulkAction' => 'VuFind\View\Helper\Root\BulkAction',
             'captcha' => 'VuFind\View\Helper\Root\Captcha',
